requirements.txt
django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
django-downloadview==2.5.0

numpy==1.26.4
pillow==12.1.0
opencv-python==4.8.1.78
onnxruntime==1.24.1

pandas==3.0.0
openpyxl==3.1.5
pdfplumber==0.11.9
python-docx==1.2.0
pywin32==311
